Extract event dropdown to separate component
